Move non-language-server code out of activate().  Enable transition from IntelliSenseEngine Disabled to enabled, without requiring reload.
